+++ src/app/cart/cart.config.ts
@@ -1,0 +1,6 @@
+import { InjectionToken } from '@angular/core';
+
+export const CartAPI = new InjectionToken<string>('CartAPI', {
+  providedIn: 'root',
+  factory: () => 'http://localhost:3000/cart'
+});