+++ src/app/cart/models/cart-item.model.ts
@@ -1,5 +1,6 @@
 export default class CartItemModel {
   constructor(
+    public id: number,
     public name: string,
     public price: number,
     public quantity: number,