+++ src/app/data/products.ts
@@ -1,5 +1,6 @@
-export const products = [
+export const default_products = [
   {
+    "id": 1,
     "name": "Maria",
     "description": "Cookies",
     "price": 15,
@@ -7,6 +8,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 2,
     "name": "Beef",
     "description": "Local farm production",
     "price": 200,
@@ -14,6 +16,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 3,
     "name": "10 good eggs",
     "description": "Eggs",
     "price": 70,
@@ -21,6 +24,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 4,
     "name": "Rosinka",
     "description": "Lemonade",
     "price": 20,
@@ -28,6 +32,7 @@ export const products = [
     "isAvailable": false
   },
   {
+    "id": 5,
     "name": "Toilet Duck",
     "description": "Cleaning Detergent",
     "price": 80,
@@ -35,6 +40,7 @@ export const products = [
     "isAvailable": false
   },
   {
+    "id": 6,
     "name": "Persil",
     "description": "Washing Detergent",
     "price": 130,
@@ -42,6 +48,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 7,
     "name": "Moet Chandon",
     "description": "Brut Imperial",
     "price": 1200,
@@ -49,6 +56,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 8,
     "name": "Minigrill",
     "description": "Fine Tosts",
     "price": 38,
@@ -56,6 +64,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 9,
     "name": "Monobank card",
     "description": "Starting package",
     "price": 50,
@@ -63,6 +72,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 10,
     "name": "Kiyvstar simcard",
     "description": "Starting package",
     "price": 100,
@@ -70,6 +80,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 11,
     "name": "Vodafone simcard",
     "description": "Starting package",
     "price": 100,
@@ -77,6 +88,7 @@ export const products = [
     "isAvailable": true
   },
   {
+    "id": 12,
     "name": "Life simcard",
     "description": "Starting package",
     "price": 100,