+++ src/app/products/services/products.service.ts
@@ -1,17 +1,35 @@
 import { Injectable } from '@angular/core';
 import ProductModel from '../models/product.model';
-import {products} from '../../data/products'
+import { BehaviorSubject, Observable } from 'rxjs';
+import * as data from '../../data/products.json'
+import {default_products} from '../../data/products'
 
 @Injectable({
   providedIn: 'root'
 })
-export class ProductsService {
+export class ProductsService { 
 
-  constructor() { }
+  constructor() {
+  }
 
-  getProducts(): Promise<ProductModel[]> {
-    return new Promise ((resolve)=>{
-      resolve (products)   
-    }).catch(error => error) as Promise<ProductModel[]>
+  //promise approach
+  // getProducts(): Promise<ProductModel[]> {
+  //   return new Promise ((resolve)=>{
+  //     resolve (default_products)   
+  //   }).catch(error => error) as Promise<ProductModel[]>
+  // }
+
+  //pure BehaviourSubject approach
+  // getProducts(): BehaviorSubject<ProductModel[]> {
+  //   const default_products_json = data 
+
+  //   return new BehaviorSubject<ProductModel[]>(default_products_json.items as ProductModel[])
+  // }
+
+  //Observable via BehaviourSubject
+  getProducts(): Observable<ProductModel[]> {
+    const default_products_json = data 
+
+    return new BehaviorSubject<ProductModel[]>(default_products_json.items as ProductModel[])
   }
 }